Create HtmlString to represent an escaped String.
While writing the server api (and using apache as a reference), I had
the idea to shake up the api a bit. Previously, server.escape was used
to escape a String so that server.write_raw could send it (or you
could instead do server.write which was a shorthand for both).

I've added HtmlString as a class, which takes a String in the
constructor. The resulting HtmlString thus presents a type-safe way of
knowing that a String has already been encoded (you just need to make
sure it isn't doubly encoded).

I've changed server.write_literal so that it no longer verifies that
the source is an interned literal. I realized that a mock server api
written in pure Lily wouldn't be able to query interned status without
adding builtin api. I don't want to do that.

server.write_raw is now server.write_unsafe. Both write_raw and
write_literal do the same action, and that's intentional. The goal is
to guide people toward using write (and auto escaping) or literal
when they have a literal. write_unsafe intentionally sounds bad. Use
it as necessary.